html/cross-origin-opener-policy/resource-popup.https.html

<!DOCTYPE html>
<html lang="en">
<title>Cross-Origin-Opener-Policy forces browsing context switch in various popup document types</title>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script>
'use strict';

/**
 * These tests create a "parent" popup window which is an HTML document with a
 * specific Cross-Origin-Opener-Policy. The parent creates a "child" popup
 * window which is a non-HTML document with a Cross-Origin-Opener-Policy of its
 * own. The HTTP `Refresh` header is used to ensure the child eventually closes
 * itself (since proper observance of COOP will prevent the parent from closing
 * the child in some cases).
 */
const make = ({t, parentCoop, resource, resourceCoop, resourceName, validate}) => {
  const bc = new BroadcastChannel(token());
  bc.onmessage = t.step_func_done(({data}) => validate(data));
  const childLocation = resource +
    '?pipe=header(Refresh,2;url=/html/cross-origin-opener-policy/resources/resource-cleanup.html)' +
    (resourceCoop ? `|header(Cross-Origin-Opener-Policy,${resourceCoop})` : '');
  const parentLocation = 'resources/resource-popup.html' +
    `?channel_name=${bc.name}` +
    `&resource=${encodeURIComponent(childLocation)}` +
    `&resource_name=${resourceName}` +
    (parentCoop ? `&pipe=header(Cross-Origin-Opener-Policy,${parentCoop})` : '');

  open(parentLocation);

  t.add_cleanup(() => {
    bc.postMessage(null);
  });
};

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resource: '/common/dummy.xml',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'xml document - parent COOP: unsafe-none; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/common/dummy.xml',
    resourceCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'xml document - parent COOP: same-origin; child COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/common/dummy.xml',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'xml document - parent COOP: same-origin; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resource: '/images/red.png',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'image document - parent COOP: unsafe-none; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/images/red.png',
    resourceCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'image document - parent COOP: same-origin; child COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/images/red.png',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'image document - parent COOP: same-origin; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resource: '/common/text-plain.txt',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'text document - parent COOP: unsafe-none; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/common/text-plain.txt',
    resourceCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'text document - parent COOP: same-origin; child COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/common/text-plain.txt',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'text document - parent COOP: same-origin; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resource: '/media/2x2-green.mp4',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'media document - parent COOP: unsafe-none; child COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/media/2x2-green.mp4',
    resourceCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'media document - parent COOP: same-origin; child COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resource: '/media/2x2-green.mp4',
    resourceCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'media document - parent COOP: same-origin; child COOP: same-origin');
</script>
</html>
